Week4/Day1/q2/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.35.0
sqlmodel==0.0.16
psycopg[binary]==3.1.19
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9 